refactor: renamed google auth provide component
